+++ appsettings.Development.json
@@ -14,11 +14,5 @@
     "Issuer": "https://bazaar.api",
     "Audience": "https://bazaar.clients",
     "ExpiryMinutes": 60
-  },
-  "AWS": {
-    "Region": "eu-north-1"
-  },
-  "S3Settings": {
-    "BucketName": "bazaar-backend-storage"
   }
 }
